find_spacy {spacyr}	R Documentation

Find spaCy

Description

Locate the user's version of Python for which spaCy installed.

Usage

find_spacy(model = "en", ask)

Arguments

model
name of the language model

ask
logical; if FALSE, use the first spaCy installation found; if TRUE, list available spaCy installations and prompt the user for which to use. If another (e.g. python_executable) is set, then this value will always be treated as FALSE.

Value

spacy_python

[Package spacyr version 0.9.91 Index]