spacy_finalize {spacyr}	R Documentation

Finalize spaCy

Description

While running spaCy on Python through R, a Python process is always running in the backgroud and rsession will take up a lot of memory (typically over 1.5GB). spacy_finalize() terminates the Python process and frees up the memory it was using.

Usage

spacy_finalize()

Author(s)

Akitaka Matsuo

[Package spacyr version 0.9.91 Index]