spacy_download_langmodel {spacyr}	R Documentation

Install a language model in a conda or virtual environment

Description

Installs one or more language models in a conda or virtualenv Python virtual environment as installed by spacy_install.

Usage

spacy_download_langmodel(model = "en", envname = "spacy_condaenv",
  conda = "auto")

spacy_download_langmodel_virtualenv(model = "en",
  envname = "spacy_virtualenv", virtualenv_root = NULL)

Arguments

model
name of the language model to be installed. A list of available language models and their names is available from the spaCy language models page.

envname
name of the virtual environment

conda
Path to conda executable. Default "auto" which automatically finds the path.

virtualenv_root
path to the virtualenv environment to install spaCy language model. If NULL, the default path "~/.virtualenvs" will be used.

[Package spacyr version 0.9.91 Index]