Parse a text using spaCy

The spacy_parse() function calls spaCy to both tokenize and tag the texts, and returns a data.table of the results. The function provides options on the types of tagsets (tagset_ options) either "google" or "detailed", as well as lemmatization (lemma). It provides a functionalities of dependency parsing and named entity recognition as an option. If "full_parse = TRUE" is provided, the function returns the most extensive list of the parsing results from spaCy.

spacy_parse(x, pos = TRUE, tag = FALSE, lemma = TRUE, entity = TRUE,
  dependency = FALSE, multithread = TRUE, ...)

Arguments

x
a character object, a quanteda corpus, or a TIF-compliant corpus data.frame (see https://github.com/ropensci/tif)

pos
logical whether to return universal dependency POS tagset http://universaldependencies.org/u/pos/)

tag
logical whether to return detailed part-of-speech tags, for the langage model en, it uses the OntoNotes 5 version of the Penn Treebank tag set (https://spacy.io/docs/usage/pos-tagging#pos-schemes). Annotation specifications for other available languages are available on the spaCy website (https://spacy.io/api/annotation).

lemma
logical; inlucde lemmatized tokens in the output (lemmatization may not work properly for non-English models)

entity
logical; if TRUE, report named entities

dependency
logical; if TRUE, analyze and return dependencies

multithread
logical; If true, the processing is parallelized using pipe functionality of spacy (https://spacy.io/api/pipe).

...
not used directly

Value

a data.frame of tokenized, parsed, and annotated tokens

Examples

spacy_initialize()

#> spaCy is already initialized

#> NULL

# See Chap 5.1 of the NLTK book, http://www.nltk.org/book/ch05.html
txt <- "And now for something completely different."
spacy_parse(txt)

#>   doc_id sentence_id token_id      token      lemma   pos entity
#> 1  text1           1        1        And        and CCONJ       
#> 2  text1           1        2        now        now   ADV       
#> 3  text1           1        3        for        for   ADP       
#> 4  text1           1        4  something  something  NOUN       
#> 5  text1           1        5 completely completely   ADV       
#> 6  text1           1        6  different  different   ADJ       
#> 7  text1           1        7          .          . PUNCT       

spacy_parse(txt, pos = TRUE, tag = TRUE)

#>   doc_id sentence_id token_id      token      lemma   pos tag entity
#> 1  text1           1        1        And        and CCONJ  CC       
#> 2  text1           1        2        now        now   ADV  RB       
#> 3  text1           1        3        for        for   ADP  IN       
#> 4  text1           1        4  something  something  NOUN  NN       
#> 5  text1           1        5 completely completely   ADV  RB       
#> 6  text1           1        6  different  different   ADJ  JJ       
#> 7  text1           1        7          .          . PUNCT   .       

spacy_parse(txt, dependency = TRUE)

#>   doc_id sentence_id token_id      token      lemma   pos head_token_id dep_rel
#> 1  text1           1        1        And        and CCONJ             3      cc
#> 2  text1           1        2        now        now   ADV             3  advmod
#> 3  text1           1        3        for        for   ADP             3    ROOT
#> 4  text1           1        4  something  something  NOUN             3    pobj
#> 5  text1           1        5 completely completely   ADV             6  advmod
#> 6  text1           1        6  different  different   ADJ             4    amod
#> 7  text1           1        7          .          . PUNCT             3   punct
#>   entity
#> 1       
#> 2       
#> 3       
#> 4       
#> 5       
#> 6       
#> 7       

txt2 <- c(doc1 = "The fast cat catches mice.\\nThe quick brown dog jumped.",
          doc2 = "This is the second document.",
          doc3 = "This is a \\\"quoted\\\" text." )
spacy_parse(txt2, entity = TRUE, dependency = TRUE)

#>    doc_id sentence_id token_id    token    lemma   pos head_token_id dep_rel
#> 1    doc1           1        1      The      the   DET             3     det
#> 2    doc1           1        2     fast     fast   ADJ             3    amod
#> 3    doc1           1        3      cat      cat  NOUN             4   nsubj
#> 4    doc1           1        4  catches    catch  VERB             4    ROOT
#> 5    doc1           1        5     mice    mouse  NOUN             4    dobj
#> 6    doc1           1        6        .        . PUNCT             4   punct
#> 7    doc1           1        7       \n       \n SPACE             6        
#> 8    doc1           2        1      The      the   DET             4     det
#> 9    doc1           2        2    quick    quick   ADJ             4    amod
#> 10   doc1           2        3    brown    brown   ADJ             4    amod
#> 11   doc1           2        4      dog      dog  NOUN             5   nsubj
#> 12   doc1           2        5   jumped     jump  VERB             5    ROOT
#> 13   doc1           2        6        .        . PUNCT             5   punct
#> 14   doc2           1        1     This     this   DET             2   nsubj
#> 15   doc2           1        2       is       be  VERB             2    ROOT
#> 16   doc2           1        3      the      the   DET             5     det
#> 17   doc2           1        4   second   second   ADJ             5    amod
#> 18   doc2           1        5 document document  NOUN             2    attr
#> 19   doc2           1        6        .        . PUNCT             2   punct
#> 20   doc3           1        1     This     this   DET             2   nsubj
#> 21   doc3           1        2       is       be  VERB             2    ROOT
#> 22   doc3           1        3        a        a   DET             7     det
#> 23   doc3           1        4        "        " PUNCT             7   punct
#> 24   doc3           1        5   quoted    quote  VERB             7    amod
#> 25   doc3           1        6        "        " PUNCT             7   punct
#> 26   doc3           1        7     text     text  NOUN             2    attr
#> 27   doc3           1        8        .        . PUNCT             2   punct
#>       entity
#> 1           
#> 2           
#> 3           
#> 4           
#> 5           
#> 6           
#> 7      GPE_B
#> 8           
#> 9           
#> 10          
#> 11          
#> 12          
#> 13          
#> 14          
#> 15          
#> 16          
#> 17 ORDINAL_B
#> 18          
#> 19          
#> 20          
#> 21          
#> 22          
#> 23          
#> 24          
#> 25          
#> 26          
#> 27